:recycle: rename a module with the same name as its parent module
