+++ src/client/document.rs
@@ -1,3 +1,3 @@
-pub mod document;
 pub mod document_contents;
 pub mod document_determs;
+pub mod document_one;
