+++ src/client/group.rs
@@ -1,2 +1,2 @@
-pub mod group;
+pub mod group_one;
 pub mod groups;
