+++ src/client/title.rs
@@ -1,2 +1,2 @@
-pub mod title;
+pub mod title_one;
 pub mod titles;
