+++ src/client/user.rs
@@ -1,2 +1,2 @@
-pub mod user;
+pub mod user_one;
 pub mod users;
